Add files via upload
These python notebooks were used to create the training data for all different kind of problems. Instead of creating different random data each time while training the model one set of data was created and saved. The saved data is called and loaded into an array while putting it into the ML model.  The motive of doing so was to keep the training data the same when we analyze the results to tune our hyperparameters.